Move MixedSpace into base
